Implemented - Increase Click Amount Upgrade
This commit adds the game's first upgrade, the ability for the player to buy an increase in their click amount. Right now each purchase of the upgrade gives the player an extra cake per click and always costs 50 cakes to purchase.

Next, implement a forever increasing price for the upgrade so that it dosen't become trivial to purchase.