Implemented - Upgrade Price Increasing
This commit adds a method which will increase the price of an upgrade by 3 times each time it is purchesed. This will work to prevent the player from abusing the purchasing of a singular upgrade to make fast progress.

Next, add a way for cakes to be gained each second irrelevant of clicking just like in Cookie Clicker.