Added - Borders + Coloured Background
This commit starts work on the actual layout of the main game, adding borders similar to Cookie Clicker which are used to seperate all the game sections. The colour of the background has also been changed for grey to dark purple as this seems more fitting for the game's theme.

Next, continue work on implementing actual game upgrades.