Implemented - Info Box
This commit adds an info box that follows the cursor around on the Y axis just like in Cookie Clicker. This will eventually provide the player with info on all upgrades and buildings when they hover over them.

Next, move the Increase CPS upgrade into the upgrade section and have the info box display the correct info.