Improved - Info Box
This commit improves the info box by a large margin, adding a title and price as well as overall better positioned text. This gives the player all the infomation they could possibly need for deciding what to purchase.

Next, have the Clicker price update in the info box when a new Clicker is bought.