Implemented - Max Clicker Building Logic
This commit implements logic for the Clicker Building to prevent the player purchasing more than the max amount of Clicker Buildings. Once max buildings is achieved the button is greyed out and the owned amount is changed to MAX.

Next, start work on the visual overhall starting with giving the borders more texture.