Added - Better Building Background
This commit adds a new background design for the different Buildings. With this the first visual overhall is complete, it's not an insane upgrade, however, each element now fits one overall theme making the game feel & look that much better.

Next, start work on adding a new building to the shop.