Implemented - Plantation Building & CPS Info in InfoBox
This commit implements functionality for a new Plantation Building and buildings CPS info to be shown, and updated when increased, in the info box that appears at the side when a building is hovered over.

Next, work on adding utilising the middle section to give a visual representation of how many plantations (and future buildings) the player currently owns.