Implemented - Plantation Upgrade 1 + Formated Price Text
This commit adds functionality for the first Plantation Upgrade and also formats the price text so it includes comma's at the relevant places. The cakes the player has is also now formated in this way.

Next, add a few more plantation upgrades + start work on promodational materials for v0.1 release.