Fixed - Broken Plantation Upgrades
This commit adds the functionality to actually apply the plantation upgrades as before I had neglected to add the actual functionality for this, instead focusing on implementing the functionality to have these upgrades appear.

Next, add a third building and possibly start thinking about the Pet System I want to have implemented.