Implemented - Pet System Section + Fixed - UI Scaling
This commit starts work on the Pet System by setting aside a section for where it will all be displayed. The UI Scaling for it has also been adjusted to work well with the rest of the UI in build.

Next, have the Pet System unlock at the correct point in progression and also start work on designing the first Pet.