Implemented - Egg Rotate/Shake Script
This commit adds a bit of hover rotation to the PetEgg sprite, this gives the PetEgg a little more life and will later allow me to implement a form of egg cracking which will atempt to give the Pet System a little more flair.

Next, design and add the first pet that will eventually be able to be hatched from the PetEgg(s).