Implemented - Egg Cracking Script
This commit removes the EggHoverShake Script and replaces it with the PetEggCrack Script which allows the player to click on the PetEgg putting more cracks in it each time. This will eventually lead to the player obtaining a Pet which gives them Cake bonuses.

Next, design and add a pet sprite.