Added - Rabbit (Pet) Sprite
This commit adds my newly designed Rabbit sprite to the project, this sprite will serve as the only pet the player can obtain for the time being. Eventually I would like to have a range of different pets that all have different raritys.

Next, implement a way for the Pet Rabbit to be hatched using the PetEgg.