Implemented - Pet Selection System
This commit implements a system for selecting the rarity of pet and which pet is actually hatched from each egg. This dosen't change anything on the player's end right now, however, it means an infinite number of pets can be added with little to no change needed for the system.

Next, implement a way to change what infomation is shown when the pet is hatched from the egg.