Implemented - 2nd Common Pet (Fish)
This commit adds my newley designed fish pet into the game, the player now has the chance to obtain the Fish pet when opening a Pet Egg. The fish gives the same bonus to CPS as the rabbit pet as right now all the common pets will give the same bonus to CPS until I decide to implement the pet rarity system.

Next, adjust how the UI scaling works so it fits more with how I feel it should work.