Implemented/Fixed - Amount of Clickers Shown + 0 Owned Bug
This commit implements functionality to show the amount of Clickers owned by the player when the save is loaded. The bug causing 0 to be shown on buildings not owned has also been fixed.

Next, implement a way for buildings that have been purchased before to be automatically made active even when the amount of Cakes held isn't enough to purchase another.