Implemented - Clear Save Button Functionality
This commit implements the functionality for the Clear Save button which allows the player to clear their saved data. After hours upon hours trying to get a scene reload to clear the data I've had to resort to forcing the player to close the game before the save data can be offically removed.

This may be re-worked down the line, however, for now this is the solution.